+++ pangea-sdk/pangea/convert_types.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	pu "github.com/pangeacyber/pangea-go/pangea-sdk/v4/internal/pangeautil"
+	pu "github.com/pangeacyber/pangea-go/pangea-sdk/v5/internal/pangeautil"
 )
 
 // Bool is a helper routine that allocates a new bool value