+++ pangea-sdk/service/file_intel/api.go
@@ -7,8 +7,8 @@ import (
 	"io"
 	"os"
 
-	"github.com/pangeacyber/pangea-go/pangea-sdk/v4/internal/request"
-	"github.com/pangeacyber/pangea-go/pangea-sdk/v4/pangea"
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/internal/request"
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 )
 
 type FileReputationRequest struct {