+++ pangea-sdk/service/redact/service.go
@@ -4,12 +4,13 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/pangeacyber/pangea-go/pangea-sdk/v4/pangea"
+	"github.com/pangeacyber/pangea-go/pangea-sdk/v5/pangea"
 )
 
 type Client interface {
 	Redact(ctx context.Context, req *TextRequest) (*pangea.PangeaResponse[TextResult], error)
 	RedactStructured(ctx context.Context, req *StructuredRequest) (*pangea.PangeaResponse[StructuredResult], error)
+	Unredact(ctx context.Context, input *UnredactRequest) (*pangea.PangeaResponse[UnredactResult], error)
 
 	// Base service methods
 	pangea.BaseServicer